Create script.js program from rock-paper-scissors
Created two functions getComputerChoice() and playRound().
Set playRound to accept two arguments, compChoice and userChoice.
Prompted user for userChoice and then did logical comparison
between compChoice and userChoice. Printed winner based on the results
of the comparison. For loop is written at the bottom of the program to
enable n rounds of rock-paper-scissors.