Add new functions to the program
Functions:
eventHandler, playRound, getRoundWinner, endMatch, updateScore,
displayWinner

Created named function eventHandler to pass the innerText value
of the button that the user selects to the playRound function.
Additionally, created the named function eventHandler to be passed to
addEventListener as a listener and then later passed to
removeEventListener once a given opponent reached a score of 5.
playRound compares playerSelection against computerSelection, calls
getRoundWinner, passing in each selection as parameters and stores the
return value. Next updateScore is called, passing in roundWinner as
a parameter. The updateScore function updates the scores dictionary
and checks if either opponent has reached a score of 5 at a given
moment. Once an opponent reaches 5, the endMatch function is called
to removeEventListener on each of the buttons and then the displayWinner
function is called.